// src/models/ICharacter.ts
export interface IResult {
    results: [],
    url: string;
    name: string;
    gender: string;
    birth_year: string;
}

export interface ICharacter {
    birth_year: string;
    eye_color: string;
    gender: string;
    hair_color: string;
    height: string;
    mass: string;
    name: string;
    skin_color: string;
}